Move delay to services-core
