+++ server/routerlicious/packages/lambdas/src/scribe/checkpointManager.ts
@@ -3,8 +3,8 @@
  * Licensed under the MIT License.
  */
 
-import { delay } from "@fluidframework/server-common-utils";
 import {
+	delay,
 	ICollection,
 	IContext,
 	isRetryEnabled,
